refactor: moves aabb and point fields to f32

+++ src/collision/collision.odin
@@ -3,12 +3,12 @@ package collision
 import rl "vendor:raylib"
 
 Point :: struct {
-	x, y: i32
+	x, y: f32
 }
 
 vector2_to_point :: proc(vec: rl.Vector2) -> Point {
 	return Point{
-		x = i32(vec[0]),
-		y = i32(vec[1]),
+		x = vec[0],
+		y = vec[1],
 	}
 }
